refactor(tui): remove obsolete package leftovers
