+++ src/presentation/ui/tui/errors.go
@@ -6,7 +6,5 @@ import (
 )
 
 var ErrUserExit = errors.New("user requested tui exit")
-var ErrBackToModeSelection = errors.New("back to mode selection requested")
 var ErrSessionClosed = bubbleTea.ErrUnifiedSessionClosed
-
 var errReconfigureRequested = errors.New("reconfigure requested")
